Introduce SID_AND_ATTRIBUTES RAII wrapper to fix memory leak (#7)
The CapabilityList variable is currently leaking Sid allocations due to the lack of a destructor in the SID_AND_ATTRIBUTES struct. This is probably not a big problem in practice, since the leak only occurs at the end of execution. Still, this is one of very few programming resources online that demonstrates how to programmatically create an AppContainer. It's therefore preferable if the code in this repo could be of high quality, since it's likely to be used as basis in other projects.